+++ packages/core/src/render3/instructions/all.ts
@@ -28,7 +28,7 @@
 export * from './attribute';
 export * from './attribute_interpolation';
 export * from './change_detection';
-export * from './container';
+export * from './template';
 export * from './storage';
 export * from './di';
 export * from './element';
